azurecloud/requirements.txt
# ============================================================================
# Enterprise RAG Platform - Python Dependencies
# ============================================================================
#
# Install with: pip install -r requirements.txt
# For development: pip install -r requirements-dev.txt
#
# ============================================================================

# -----------------------------------------------------------------------------
# Azure SDKs
# -----------------------------------------------------------------------------
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-cosmos>=4.5.1
azure-search-documents>=11.4.0
azure-ai-formrecognizer>=3.3.2
azure-functions>=1.17.0
azure-functions-durable>=1.2.4
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry>=1.2.0

# -----------------------------------------------------------------------------
# OpenAI / LLM
# -----------------------------------------------------------------------------
openai>=1.12.0
tiktoken>=0.5.2
langchain>=0.1.0
langchain-openai>=0.0.5

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
numpy>=1.26.3
pandas>=2.1.4
pydantic>=2.5.3

# -----------------------------------------------------------------------------
# Document Parsing
# -----------------------------------------------------------------------------
pypdf>=3.17.4
python-docx>=1.1.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.2
markdown>=3.5.1
openpyxl>=3.1.2

# -----------------------------------------------------------------------------
# Async / HTTP
# -----------------------------------------------------------------------------
aiohttp>=3.9.1
httpx>=0.26.0
aiofiles>=23.2.1

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.3
cachetools>=5.3.2

# -----------------------------------------------------------------------------
# Monitoring / Telemetry
# -----------------------------------------------------------------------------
opencensus-ext-azure>=1.1.13
applicationinsights>=0.11.10
